+++ src/components/EntityTypeEditor/Editor/PropertyDefinitions/index.ts
@@ -1,1 +1,1 @@
-export * from './Properties';
+export * from './PropertyDefinitions';
